backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "pycryptodome>=3.23.0",
    "pydantic-settings>=2.13.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.33",
]
